requirements.txt
nltk==3.4.5
transformers==4.2.0
scikit-learn==0.22.1
faiss-gpu==1.7.0
tqdm==4.40.0
requests
